added headers for bearer tokens so only authorized users can view BP entries, add or delete entries

+++ frontend/src/components/BPDetails.js
@@ -1,13 +1,25 @@
+import { useAuthContext } from "../hooks/useAuthContext"
 import { useBPContext } from "../hooks/useBPEntriesContext"
 import { format } from 'date-fns'
 
 const BPDetails = ({battingPractice}) => {
 
-const { dispatch }  = useBPContext()
+const { dispatch } = useBPContext()
+const { user } = useAuthContext()
+
+
 
   const handleClick = async () => {
+    if(!user) {
+      return 
+    }
+
     const response = await fetch('/api/bp-data/' + battingPractice._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers:{
+        'Authorization':`Bearer ${user.token}`,
+
+      }
     })
 
     const json = await response.json()
